Add files via upload
I tried to clean the coordinates with the clean_coordinate function, which (I think) worked! However, I then had some issues with removing those records that included land coordinates. I have written a super big function for that. I'll just quickly explain the aim of the function: the clean_coordinates function gives you those records that were flagged (so which need to be cleaned out) in a df. Therefore, I transformed my initial list of coral order into a df, then merged this df with the df from clean_coordinates. Then I ran a duplicated function on this new, combined df. However, I had to run it from the last element because otherwise it wouldn't really show me the duplicated elements in the beginning of the combined df (which correspond to the records of the list -> df ). The duplicated function gives you a logical vector (True is duplicated, false is unique). So I then transformed all the elements in the combined df to NAs where the vector gave the value TRUE. Finally, I removed the records from the combined list that correspond to the records from the clean_coordinates df. Then I removed all the NAs from the list via the na.omit function, and transformed the final df into a new list. I know this function is highly complicated and super inefficient (given the time it takes). I will probably be a little frustrated when you show me that there is an easy command that allows you to do all of this in one line BUT I am a little proud that both of my functions work at least (even if they don't quite do what they're supposed to). It's already an improvement from last time I guess.